# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CleanOps Pro - Multi-tenant SaaS API for cleaning businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, overdue]
        - name: client_id
          in: query
          schema:
            type: integer
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"

    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceRequest"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/send:
    post:
      operationId: sendInvoice
      tags: [invoices]
      summary: Send invoice to client
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Invoice sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    InvoiceLineItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unit_price:
          type: number
        total:
          type: number
      required:
        - description
        - quantity
        - unit_price
        - total
    Invoice:
      type: object
      properties:
        id:
          type: integer
        client_id:
          type: integer
        client_name:
          type: string
        job_id:
          type: integer
          nullable: true
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        line_items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        subtotal:
          type: number
        tips:
          type: number
        total:
          type: number
        created_at:
          type: string
          format: date-time
        paid_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - client_id
        - client_name
        - status
        - line_items
        - subtotal
        - tips
        - total
        - created_at
    InvoiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        stats:
          type: object
          properties:
            total_revenue:
              type: number
            total_paid:
              type: number
            total_overdue:
              type: number
            total_outstanding:
              type: number
          required:
            - total_revenue
            - total_paid
            - total_overdue
            - total_outstanding
      required:
        - data
        - total
        - page
        - limit
    CreateInvoiceRequest:
      type: object
      properties:
        client_id:
          type: integer
        job_id:
          type: integer
          nullable: true
        line_items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        tips:
          type: number
      required:
        - client_id
        - line_items
    UpdateInvoiceRequest:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        line_items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        tips:
          type: number
...